[FIX] Fix deleteGroup Api - 그룹 삭제 시 user도 삭제되는 오류 해결 - 그룹 삭제 시 user의 role은 user로 변경
related to #5
