+++ src/main/java/gachicar/gachicarserver/domain/Group.java
@@ -20,10 +20,10 @@ public class Group {
     @Setter
     private String desc;
 
-    @OneToMany(mappedBy = "group", cascade = CascadeType.ALL)
+    @OneToMany(mappedBy = "group")
     private List<User> memberList = new ArrayList<>();
 
-    @ManyToOne(fetch = FetchType.LAZY)
+    @OneToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "manager_id")
     private User manager; // 그룹장
