+++ src/main/java/gachicar/gachicarserver/service/UserService.java
@@ -3,7 +3,10 @@ package gachicar.gachicarserver.service;
 import gachicar.gachicarserver.domain.Group;
 import gachicar.gachicarserver.domain.Role;
 import gachicar.gachicarserver.domain.User;
+import gachicar.gachicarserver.dto.requestDto.DeleteGroupRequestDto;
 import gachicar.gachicarserver.dto.requestDto.UpdateUserNicknameRequestDto;
+import gachicar.gachicarserver.exception.ApiErrorException;
+import gachicar.gachicarserver.exception.ApiErrorStatus;
 import gachicar.gachicarserver.exception.AuthErrorException;
 import gachicar.gachicarserver.exception.AuthErrorStatus;
 import gachicar.gachicarserver.repository.UserRepository;
@@ -12,6 +15,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 @Transactional(readOnly = true)
@@ -52,5 +56,25 @@ public class UserService {
     public void updateGroup(User user, Group group) {
         user.setGroup(group);
         user.setRole(Role.MANAGER);
+    }
+
+    @Transactional
+    /* 그룹 삭제 */
+    public void deleteGroup(User user, DeleteGroupRequestDto requestDto) {
+        Group group = user.getGroup();
+
+        if (group != null) {
+            if (Objects.equals(group.getGroupId(), requestDto.getDeleteId())) {
+                // 사용자가 그룹장인지 확인
+                if (group.getManager() == user) {
+                    user.setRole(Role.USER);
+                    user.setGroup(null);
+                } else {
+                    throw new ApiErrorException(ApiErrorStatus.NOT_MANAGER);
+                }
+            }
+        } else {
+            throw new ApiErrorException(ApiErrorStatus.NOT_HAVE_GROUP);
+        }
     }
 }
